# Process events synchronously