This is a follow up to #75.

Previously, each event was processed in its own goroutine. This lead to
problematic order of event processing (e.g, node add|remove|modify where the
modify is trying to modify something that doesn't exist). It also lead to
goroutines waiting for things that would never subsequently happen.

Two significant changes in this PR:

* Do not run updates and event handlers each in their own goroutine, but
  instead run them in one goroutine in the order they came in. This
  avoids the race described in #75.

* Split the process queues into their own goroutines. This is necessary
  so that event handling can run commands. If all the queues are
  processed from the same goroutine, then when for exapmle when
  `documentUpdated()` runs,  it tries to run a `dom.GetDocument()`,
  which would deadlock because documentUpdated() is occupying the
  handler goroutine.

This patch does not represent a complete fix, because I believe it may make
a lot of locking and other concurrent machinary unnecessary. I don't have
time to explore that right now, but I don't think it's harmful as far as I
can tell.

I think there should in principle be a follow up PR to remove now-unnecessary
locks and waitgroups, but I'm not in a position to do those changes myself just
at the moment.