@pwaller 2017-06-21
/cc @rjeczalik 

There is also this [patch](https://github.com/rjeczalik/chromedp/commit/1d3ee5b2625b76a2e3f77b8ec1be7512cfb8d2cb) to what I've done to ignore waits on empty nodes. I haven't hit that case myself and the commit has a comment stating that it might not be the best way to handle it. It would be useful to have a reproducer for it so that we can make a better fix for it.

I neglected to mention in the PR body 00b69aa which ignores EventFrame state changes. This is necessary because those events seem to genuinely refer to FrameIDs which don't yet (and might never) have a corresponding `cdf.Frame` object associated with them. So for now I simply drop those, since it seems not possible to use those state changes correctly in the current implementation.