@rjeczalik 2017-06-21
>  It would be useful to have a reproducer for it so that we can make a better fix for it.

I think the bug manifests itself when we cdp.Navigate to a different url while current frame is still being loaded - it may somehow invalidate old nodes thus 0 received via cdp events. I did not debug it, this is just my wild guess.