@pwaller 2017-06-21
Are you able to come up with some minimal setup which occasionally triggers ityou can share?