@rjeczalik 2017-06-21
@pwaller If I got couple of spare minutes I'm going to reproduce this and create separate issue/PR. For now just sharing observations if it happens someone will get around solving this problem faster.