@pwaller 2017-06-30
I think I've found the root cause of the waiting on empty nodes. It's actually because `documentUpdated` puts the nodes on the wrong frame. I've struggled to find a reliable simple reproducer - loading a Google document does the trick.

I think the underlying problem here is with the way that `h.cur` works. The problem is that we want to update it at the point of the navigation, but at that point we don't yet have a `*cdp.Frame`. This only comes later.

It might be better if the `h.cur` mechanism didn't store the frame object but just the ID - and the actual user of the frame then calls `WaitFrame`. I'll give this a try if I get a moment.

```
2017/06/30 20:22:28 cmd 9/"Page.navigate"
2017/06/30 20:22:29 Got result for 9/""
2017/06/30 20:22:29 Event *page.EventFrameStartedLoading
2017/06/30 20:22:29 SetActive waiting for frame "19500.1" // caused by navigation
2017/06/30 20:22:29 Event *page.EventFrameNavigated
2017/06/30 20:22:29 Event *dom.EventDocumentUpdated
2017/06/30 20:22:29 WaitFrame(id="") = 0xc420086140
2017/06/30 20:22:29 documentUpdated operating on 0xc420086140 starts
2017/06/30 20:22:29 cmd 10/"DOM.getDocument"
2017/06/30 20:22:29 WaitFrame(id="19500.1") = 0xc4224d6320 // SetActive's wait on the frame finally completes.
2017/06/30 20:22:29 SetActive sets h.cur = 0xc4224d6320 // caused by navigation
2017/06/30 20:22:29 Got result for 10/""
2017/06/30 20:22:29 documentUpdated operating on 0xc420086140 completes // Old h.cur
2017/06/30 20:22:29 Event *dom.EventChildNodeCountUpdated
2017/06/30 20:22:29 WaitFrame(id="") = 0xc4224d6320 // New h.cur
2017/06/30 20:22:30 WaitNode spinning on 0xc4224d6320
```