@kenshaw 2017-07-01
So, we can't really make this synchronous. However, there absolutely is a problem with the implementation, and I think you're partially correct that it should store the frame ID and then essentially wait until that frame i present. What I'm going to do is split out the code from the current handler, into a separate manager type. I don't have a fully formed idea yet on this, but it'll have significant implications for the API in general. I'll play around with it, and compare some different approaches / styles to see what the best solution would be.