@pwaller 2017-07-01
> So, we can't really make this synchronous.

Could you elaborate on the reasoning? From my research I disagree (to a degree). This statement by itself makes it hard to give any useful feedback.  I also fear we may talk cross purposes unless I can better understand your motivations.

There are some things that can be done asynchronously, but I think specifically that processing every incoming event in an arbitrary order is  complicating things and hiding several bugs such as the one in my comment above. These bugs make it quite hard to get started with ChromeDP - see for example #80, #75, [this comment](https://github.com/knq/chromedp/issues/82#issuecomment-312282046). I also believe that there are often goroutines spinning waiting for elements which will never appear because of these issues.

I believe that certain heavy lifting can be done in parallel (if that is the motivation - and if it is 
 worth it) while still doing the event processing in-order, with some care. I think the net result could end up being easier to reason about than the current approach, and possibly faster.