@derlaft 2017-08-13
@knq, whats your status about it? Without this patch I get tons of that errors and lots of seconds of lags.