@t-blake 2017-08-16
I have been using this patch for several days and it has eliminated the lags and missing node errors that I was experiencing with the master branch. 
Many thanks @pwaller