@jcrussell 2017-11-29
I made some cleanups and merged to master: https://github.com/jcrussell/chromedp/commit/cf34da2947650178ca6d1550c3c02de9b6750b60.

Seems to be working great!