@bcspragu 2017-12-05
I'm currently evaluating knq/chromedp and tebeka/selenium for some browser automation tasks. Without this pull request, a basic 'Hello World' style example takes ~30s in chromedp and ~2s in selenium, despite chromedp being the 'lightweight' approach. 

The example in ChromeDP:
```go
err = c.Run(ctx, cdp.Tasks{
  cdp.Navigate("http://webscraper.io/test-sites/e-commerce/allinone"),
  cdp.Click("//a[text()[contains(., 'Computers')]]", cdp.NodeVisible),
  cdp.Click("//a[text()[contains(., 'Laptops')]]", cdp.NodeVisible),
  cdp.Nodes(".thumbnail .title", &titles, cdp.ByQueryAll),
})
```

And now in Selenium:

```go
if err := wd.Get("http://webscraper.io/test-sites/e-commerce/allinone"); err != nil {
  return nil, err
}

el, err := wd.FindElement(selenium.ByXPATH, "//a[contains(text(), 'Computers')]")
if err != nil {
  return nil, err
}

if err := el.Click(); err != nil {
  return nil, err
}

el, err = wd.FindElement(selenium.ByXPATH, "//a[contains(text(), 'Laptops')]")
if err != nil {
  return nil, err
}

if err := el.Click(); err != nil {
  return nil, err
}

els, err := wd.FindElements(selenium.ByCSSSelector, ".thumbnail .title")
if err != nil {
  return nil, err
}
```

Without this patch:

```
$ go test --bench=. --benchtime=30s

goos: linux
goarch: amd64
pkg: github.com/bcspragu/auto-test
BenchmarkChromeDP-12                   1        32088692546 ns/op
BenchmarkSelenium-12                  20        2212985810 ns/op
PASS
ok      github.com/bcspragu/auto-test   80.671s
```

With this pull request applied, the results are more even:

```
$ go test --bench=. --benchtime=30s

goos: linux
goarch: amd64
pkg: github.com/bcspragu/auto-test
BenchmarkChromeDP-12                  20        1993346253 ns/op
BenchmarkSelenium-12                  20        2238880901 ns/op
PASS
ok      github.com/bcspragu/auto-test   91.212s
```

I'd like to use (and contribute to) the ChromeDP library, because the API is far easier to work with and doesn't require Java/Selenium, but being 15x slower is a non-starter. Is there some way to reconcile a fix here?