@stephencheng 2017-12-17
@bcspragu it's just a matter of time this fix will be merged back. I am in the same boat and currently use the temp fix, till then, it will only take millseconds for the job