@stephencheng 2017-12-29
@kenshaw seems this pull request was not merged back to master branch.

Understand that you have the project split. What's your plan or thought to adopt this pull request? I wish it could be merged sooner than late as it helps the performance